requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.5.6
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
blis==0.7.4
cachetools==4.2.4
catalogue==2.0.4
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
cycler==0.10.0
cymem==2.0.5
datasets==2.4.0
dill==0.3.5.1
emoji==1.4.2
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.7.1
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-bigquery==2.29.0
google-cloud-core==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.12.0
isort==5.8.0
Jinja2==3.0.1
joblib==1.1.0
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.5
nltk==3.6.2
numpy==1.21.6
oauthlib==3.1.0
packaging==21.3
pandas==1.3.5
pathy==0.5.2
Pillow==9.2.0
plotly==5.10.0
preshed==3.0.5
proto-plus==1.19.7
protobuf==4.21.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.7.4
pylint==2.8.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.0
responses==0.18.0
rsa==4.7.2
scikit-learn==1.0.2
scipy==1.7.3
searchtweets-v2==1.1.1
sentencepiece==0.1.97
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
tenacity==8.1.0
thinc==8.0.3
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.60.0
transformers==4.21.2
tweet-parser==1.13.2
typed-ast==1.4.3
typer==0.3.2
typing_extensions==4.3.0
urllib3==1.26.12
wasabi==0.8.2
wordcloud==1.8.1
wrapt==1.12.1
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
